requirements.txt
pandas
scikit-learn
apache-airflow
joblib